Add --enable-cleartext-plugin option (#32)
* Add --enable-cleartext-plugin option

This changeset adds the `--enable-cleartext-plugin` param to allow
passing in cleartext passwords.

This is passed to the adapter's `allow_local_infile` field.

A workaround, without this configuration param, is setting the envvar
`LIBMYSQL_ENABLE_CLEARTEXT_PLUGIN`. However, the plugin config param is
preferable because it allows configuration on a per-connection basis.

* Run checks